fix: wip refactor onselect
